requirements.txt
numpy
gradio
ultralytics
supervision
matplotlib
tqdm
transformers
pillow
